code P. & update/delate comment

+++ delete-comment.php
@@ -1,18 +1,27 @@
 <?php
+require_once ('librairie/tools.php');
 
 $id = null;
+$idC = null;
+
 if (!empty($_GET['id']) && ctype_digit($_GET['id']))
 {
     $id = $_GET['id'];
 }
+if (!empty($_GET['idC']) && ctype_digit($_GET['idC']))
+{
+    $idC = $_GET['idC'];
+}
 
-require_once ('pdo.php');
+if ($idC)
+{
+    require_once ('pdo.php');
 
-$requestDelete = $pdo->prepare('DELETE FROM comments WHERE id= :id');
-$requestDelete->execute([
-    "id"=>$id
-]);
+    $requestDelete = $pdo->prepare('DELETE FROM comments WHERE id= :id');
+    $requestDelete->execute([
+        "id"=>$idC
+    ]);
 
-header('Location: index.php');
-
+    redirect("post.php?id=${id}");
+}
 ?>
